other_services/ml_services/requirements.txt
aiokafka==0.12.0
gmqtt==0.6.16
google-generativeai==0.8.3
colorlog==6.8.2
python-dotenv==1.0.1